move notification to neoscannode
