+++ apps/neoscan_node/lib/neoscan_node.ex
@@ -1,6 +1,7 @@
 defmodule NeoscanNode do
   @moduledoc false
   alias NeoscanNode.Worker
+  alias NeoscanNode.Notifications
 
   def get_nodes do
     Worker.get_nodes()
@@ -12,5 +13,9 @@ defmodule NeoscanNode do
 
   def get_data do
     Worker.get_data()
+  end
+
+  def add_notifications(block, height) do
+    Notifications.add_notifications(block, height)
   end
 end
